% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_datasets.R
\name{check_datasets}
\alias{check_datasets}
\title{Perform some POEM specific checks on datasets}
\usage{
check_datasets(dataset_list)
}
\arguments{
\item{dataset_list}{list of datasets formatted according to the requirements of coloc.abf with the following differences
pos, and chr are additionally required.
snp is ignored as it is determined by pos.
For top SNP datasets, R2 and MAF and gds_file and subset are optional, if supplied as arguments to the function.
For full summary statistic datasets, MAF is required.}
}
\description{
Perform some POEM specific checks on datasets
}
\examples{
dataset_full <- list(pos = c(1, 2, 3, 4, 5), MAF = c(0.14, 0.15, 0.25, 0.2, 0.4), N=1000, type ="quant", pvalues = c(2 * 10^-8, 4 * 10^-8, 2 * 10^-4, 0.6, 0.03), chr= "Z")
dataset_top_SNP <- list(pos = 2, N= 10000, s =0.5, type="cc", pvalues = 10^-9, chr = "Z")
dataset_top_SNP_no_chr <- list(pos = 2, N= 10000, s =0.5, type="cc", pvalues = 10^-9)
dataset_top_SNP_different_chr <- list(pos = 2, N= 10000, s =0.5, type="cc", pvalues = 10^-9, chr = "Y")
dataset_full_zero_p <- list(pos = c(1, 2, 3, 4, 5), MAF = c(0.14, 0.15, 0.25, 0.2, 0.4), N=1000, type ="quant", pvalues = rep(0,5), chr= "Z")
# Should pass
check_datasets(list(dataset_full, dataset_top_SNP))
# Should fail
try(check_datasets(list(dataset_full, dataset_top_SNP, dataset_top_SNP_no_chr)))
try(check_datasets(list(dataset_full, dataset_top_SNP, dataset_top_SNP_different_chr)))
try(check_datasets(list(dataset_full, dataset_top_SNP, dataset_full_zero_p)))
}
