% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_condition.R
\name{check_input_condition}
\alias{check_input_condition}
\title{see if input is from function or dataset}
\usage{
check_input_condition(dataset_input, function_input)
}
\arguments{
\item{dataset_input}{the corresponding item from the dataset}

\item{function_input}{the corresponding item from the function}
}
\value{
'both' if in both dataset and function 'function' if function only, 'dataset' if dataset only, 'none' if neither
}
\description{
see if input is from function or dataset
}
