% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_R2_from_geno_df.R
\name{compute_R2_MAF_from_geno_df}
\alias{compute_R2_MAF_from_geno_df}
\title{Given output from getSNP, compute the R2 and MAF}
\usage{
compute_R2_MAF_from_geno_df(gdsout, dataset)
}
\arguments{
\item{gdsout}{output from getSNP}

\item{dataset}{must have top_pos giving the top SNP position in dataset}
}
\description{
Given output from getSNP, compute the R2 and MAF
}
