% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_imputation_class.R
\name{detect_imputation_class}
\alias{detect_imputation_class}
\title{Determine whether dataset has full or top SNP statistics}
\usage{
detect_imputation_class(dataset)
}
\arguments{
\item{dataset}{POEMColoc formated dataset}
}
\value{
"top" or "all'
}
\description{
Determine whether dataset has full or top SNP statistics
}
\examples{
# p-value input
dataset_full <- list(pos = c(1, 2, 3, 4, 5), MAF = c(0.14, 0.15, 0.25, 0.2, 0.4), N=1000, type ="quant", pvalues = c(2 * 10^-8, 4 * 10^-8, 2 * 10^-4, 0.6, 0.03), chr= "Z")
dataset_top_SNP <- list(pos = 2, N= 10000, s =0.5, type="cc", pvalues = 10^-9, chr = "Z")
detect_imputation_class(dataset_full)
detect_imputation_class(dataset_top_SNP)
# beta-varbeta input
dataset_full_beta_varbeta = dataset_full
dataset_full_beta_varbeta$pvalues = NULL
dataset_full_beta_varbeta$beta = c(0.5, 0.5, 0.4, 0.5, 0.26)
dataset_full_beta_varbeta$varbeta = c(0.008, 0.008, 0.012, 0.01, 0.015)
dataset_full_beta_varbeta$sdY = 1
dataset_top_SNP_beta <- list(pos = 2, N= 10000, s =0.5, type="cc", beta = 2, varbeta = 0.05, chr = "Z")
detect_imputation_class(dataset_full_beta_varbeta)
detect_imputation_class(dataset = dataset_top_SNP_beta)
}
