% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_list_condition.R
\name{detect_list_condition}
\alias{detect_list_condition}
\title{Title}
\usage{
detect_list_condition(dataset)
}
\arguments{
\item{dataset}{POEMColoc input dataset}
}
\value{
TRUE if a list of datasets, FALSE if a single dataset
}
\description{
Title
}
