% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_POEMColoc_input.R
\name{format_POEMColoc_input}
\alias{format_POEMColoc_input}
\title{Given arguments to POEMColoc, convert input to standardized format for imputation step.}
\usage{
format_POEMColoc_input(
  dataset_list,
  gds_file,
  R2,
  MAF,
  subset,
  window_size,
  get_dosage_fn = get_dosage_alt,
  flexible_input = TRUE
)
}
\arguments{
\item{dataset_list}{list of datasets formatted according to the requirements of coloc.abf with the following differences
pos, and chr are additionally required.
snp is ignored as it is determined by pos.
For top SNP datasets, R2 and MAF and gds_file and subset are optional, if supplied as arguments to the function.
For full summary statistic datasets, MAF is required.}

\item{gds_file}{path to a gds file for the reference panel.  Can be NULL if specifying R2 and MAF, or if supplying within dataset.}

\item{R2}{A matrix of R2 LD between SNP positions.  This should contain LD information for top SNPs in rows and LD for positions to be imputed in columns.  Can be NULL if specifying gds_file or if included in individual datasets.
Column and row names should be position.}

\item{MAF}{optional input of minor allele frequencies.
only used for top SNP datasets
names should be position}

\item{subset}{optional parameter that is the path to a file giving the names of samples to include as formatted in the gds file.  each line should be the name of a sample.}

\item{window_size}{if NULL, window size is determined to cover all positions in dataset_list
if specified, covers all positions in dataset list plus window_size around top SNP dataset}

\item{get_dosage_fn}{a function taking gds file connection and returning
a matrix with the genotype or imputed dosage at each position.  The function
should not perform any filtering or reordering of variants.  rows must correspond
to individuals and columns to positions.}
}
\value{
POEM_input format dataset
Specifically, this is a coloc format dataset with the following additions
pos length L (where L is length of pvalues or beta)
chr length 1
snp length L equal to as.character(pos) is mandatory
imputation class either all or top
top_pos giving the position of the top p-value
if imputation class is top
-R2 which is either a named vector R2 with names being snp of length M or NA
-MAF which is either a named vector with names being snp of length P or NA
if imputation class is 'all'
-vector MAF of length L
}
\description{
Given arguments to POEMColoc, convert input to standardized format for imputation step.
}
