% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_priors.R
\name{format_priors}
\alias{format_priors}
\title{format_priors}
\usage{
format_priors(dataset_list, config, p1, p12)
}
\arguments{
\item{dataset_list}{list of datasets to be analyzed of length S}

\item{config}{list of vectors of indices 1:S of length R}

\item{p1}{vector of probability of length S or 1}

\item{p12}{vector of probability of length R or 1}
}
\value{
list with elements
p1 vector of probability of length S
p12 vector of probability of length R
}
\description{
format_priors
}
