% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imputed_dosage.R
\name{get_imputed_dosage}
\alias{get_imputed_dosage}
\title{wrapper for seqGetData}
\usage{
get_imputed_dosage(gds)
}
\arguments{
\item{gds}{gds file connection}
}
\value{
matrix with rows corresponding to samples and columns to positions
}
\description{
wrapper for seqGetData
}
