% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSNP.R
\name{getSNP}
\alias{getSNP}
\title{Get genotypes in region}
\usage{
getSNP(
  gds_file,
  chr,
  start,
  end,
  subset = NULL,
  get_dosage_fn = get_dosage_alt
)
}
\arguments{
\item{gds_file}{Path to gds file}

\item{chr}{Chromesome id}

\item{start}{start position of region on chromosome}

\item{end}{end position of region on chromosome}

\item{subset}{optional samples file for input to bcftools}

\item{get_dosage_fn}{a function taking gds file connection and returning
a matrix with the genotype or imputed dosage at each position.  The function
should not perform any filtering or reordering of variants.  rows must correspond
to individuals and columns to positions.}
}
\value{
A data.frame of alleles
}
\description{
Get genotypes in region
}
\examples{

}
