% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputed_coloc_input.R
\name{imputed_coloc_input}
\alias{imputed_coloc_input}
\title{Imputed coloc input}
\usage{
imputed_coloc_input(dataset)
}
\arguments{
\item{dataset}{POEM_input format dataset
Specifically, this is a coloc format dataset with the following additions
pos length L (where L is length of pvalues or beta)
chr length 1
snp length L equal to as.character(pos) is mandatory
imputation class either all or top
top_pos giving the position of the top p-value
if imputation class is top
-R2 which is either a named vector R2 with names being snp of length M or NA
-MAF which is either a named vector with names being snp of length P or NA
if imputation class is 'all'
-vector MAF of length L}

\item{gds_file}{path to a gds file for the reference panel.  Can be NULL if specifying R2 and MAF, or if supplying within dataset.}

\item{subset}{optional parameter that is the path to a file giving the names of samples to include as formatted in the gds file or vcf from which it came.  each line should be the name of a sample.}

\item{R2}{A matrix of R2 LD between SNP positions.  This should contain LD information for top SNPs in rows and LD for positions to be imputed in columns.  Can be NULL if specifying gds_file or if included in individual datasets.
Column and row names should be position.}

\item{MAF}{optional input of minor allele frequencies.
Will also be used for full summary statistic datasets if not specified within the dataset.}

\item{window_size}{It is required if there are two top SNP datasets.
Default NULL will impute all positions present in the full datasets.
It should not be NULL if all datasets are top SNP datasets as this will impute everything between the two top SNPs only.
Otherwise, for each pairwise colocalization, only SNPs within window size of one of their top SNPs will be included, including all SNPs between them if disjoint.}
}
\value{
POEM_output format dataset
Specifically this is a coloc format dataset with the following changes
top_pos giving the position of the top p-value
pos vector giving positions
chr giving chromosome
snp is mandatory
MAF is mandatory
imputation class that is either all or top
Unlike coloc, length of pvalues or beta L is permitted to be zero
positions included will be those in R2 and MAF that are not NA
}
\description{
Imputed coloc input
}
\examples{
R2 = c(0.9, 1, 0.5, 0, 0.2)
names(R2) <- 1:5
MAF = c(0.1, 0.15, 0.3, 0.2, 0.4)
names(MAF) <- 1:5
# Top snp example
dataset_top_SNP <- list(pos = 2, MAF = MAF, R2=R2, N= 10000, s =0.5, type="cc", pvalues = 10^-9, chr = "Z", imputation_class = "top", top_pos=2, snp="2")
imputed_coloc_input(dataset = dataset_top_SNP)
# Full example
dataset_full <- list(pos = c(1, 2, 3, 4, 5), MAF = c(0.14, 0.15, 0.25, 0.2, 0.4), N=1000, type ="quant", pvalues = c(2 * 10^-8, 4 * 10^-8, 2 * 10^-4, 0.6, 0.03), chr= "Z", imputation_class = "all", top_pos = 1, snp=as.character(1:5))
imputed_coloc_input(dataset = dataset_full)
# Top snp example with beta
dataset_top_SNP_beta <- list(pos = 2, snp = "2", MAF = MAF, R2=R2, N= 10000, s =0.5, type="cc", beta = 2, varbeta = 0.05, chr = "Z", imputation_class = "top", top_pos=2)
imputed_coloc_input(dataset = dataset_top_SNP_beta)
# Full dataset example with beta
dataset_full_beta_varbeta = dataset_full
dataset_full_beta_varbeta$pvalues = NULL
dataset_full_beta_varbeta$beta = c(0.5, 0.5, 0.4, 0.5, 0.26)
dataset_full_beta_varbeta$varbeta = c(0.008, 0.008, 0.012, 0.01, 0.015)
dataset_full_beta_varbeta$sdY = 1
imputed_coloc_input(dataset = dataset_full_beta_varbeta)
# Top SNP dataset with NA R2
dataset_top_SNP_NA <- list(pos = 2, MAF = MAF, R2=NA, N= 10000, s =0.5, type="cc", pvalues = 10^-9, chr = "Z", imputation_class = "top", top_pos=2)
imputed_coloc_input(dataset = dataset_top_SNP_NA)
# Top SNP dataset with missing MAF positions
dataset_top_SNP_beta2 <- list(pos = 2, snp = "2", MAF = MAF[-1], R2=R2, N= 10000, s =0.5, type="cc", beta = 2, varbeta = 0.05, chr = "Z", imputation_class = "top", top_pos=2)
imputed_coloc_input(dataset = dataset_top_SNP_beta2)
}
