% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POEMColoc.R
\name{POEMColoc}
\alias{POEMColoc}
\title{Perform coloc analysis using imputed p-values}
\usage{
POEMColoc(
  dataset,
  dataset_list,
  gds_file = NULL,
  R2 = NULL,
  MAF = NULL,
  subset = NULL,
  window_size = NULL,
  min_MAF = 0,
  p1 = 10^-4,
  p2 = 10^-4,
  p12 = 10^-5,
  dosage_method = "dosage_alt"
)
}
\arguments{
\item{dataset}{single dataset formatted according to the requirements of coloc.abf with the following differences
pos, and chr are additionally required.
snp is ignored as it is determined by pos.}

\item{dataset_list}{list of datasets formatted according to the requirements of coloc.abf with the following differences
pos, and chr are additionally required.
snp is ignored as it is determined by pos.}

\item{gds_file}{path to a gds file for the reference panel.  Can be NULL if specifying R2 and MAF.}

\item{R2}{A matrix of R2 LD between SNP positions.  This should contain LD information for top SNPs in rows and LD for positions to be imputed in columns.  Can be NULL if specifying gds_file.
Column and row names should be position.}

\item{MAF}{optional input of minor allele frequencies.
only used for top SNP datasets
names should be position}

\item{subset}{optional parameter that is the path to a file giving the names of samples to include as formatted in the gds file.  each line should be the name of a sample.}

\item{window_size}{It is required if there are two top SNP datasets.
Default NULL will impute all positions present in the full datasets.
It should not be NULL if all datasets are top SNP datasets as this will impute everything between the two top SNPs only.
Otherwise, for each pairwise colocalization, only SNPs within window size of one of their top SNPs will be included, including all SNPs between them if disjoint.}

\item{min_MAF}{Remove any site with minor allele frequency less than min_MAF.  Setting to zero means no filtering.}

\item{p1}{prior probability a SNP is associated with trait in dataset, default 1e-4.}

\item{p2}{prior probability a SNP is associated with trait t, where t is the index in dataset list, default 1e-4.  Either a single number or a vector of length same as dataset_list.}

\item{p12}{prior probability a SNP is associated with both traits, default 1e-5.  Either a single value or of the same length as dataset_list.}

\item{dosage_method}{Method to compute dosage if gds_file input mode is used.  one of 'dosage', 'dosage_alt' or 'dosage_imputed', see details.}
}
\value{
list of coloc.abf results.
}
\description{
Perform coloc analysis using imputed p-values
}
\details{
dataset and (elements of) dataset_list follow the format for coloc.abf in the coloc R package, except snp is not optional and additional elements pos and chr giving genome position and chromosome id are required.
Additionally, MAF need not be supplied for datasets to be imputed because this will be obtained elsewhere.
One or both of the two datasets can be in top SNP format.  snp will be overridden with pos so it is expected that all sources will use the same coordinate system.
Multiple alternative allele per site are collapsed.
If you have a vcf file first convert to gds using SeqArray:::seqVCF2GDS(your_vcf, your_gds)
prior parameters p1, p2, p12 are supplied to the coloc.abf function
POEMColoc creates imputed datasets for input to the coloc.abf function, refer to the coloc.abf documentation from the coloc R package for more information on this method and its inputs
pos and chr are additional requirements of POEMColoc in order to recognize each position in R2 and MAF, or in the gds file in the panel.
There are different ways to run POEMColoc.
POEMColoc can be run with R2 and MAF supplied as function arguments.  Note in this case R2 must be a matrix with both rownames including any top SNP positions to be imputed from and column names being positions of any SNPs to impute, including the top SNP itself.
Datasets consisting of more than one SNP will not be imputed.
Alternatively, R2 and MAF can be computed from a gds file and optional file containing subset of samples to use, which can be created from a vcf file (see above).
Note that user-supplied R2 and MAF will take priority over gds file.
dosage_method is designed to support different gds input possibilities.
'dosage' will compute reference allele dosage, 'dosage_alt' will compute dosages of alternative allele, imputed_dosage will use 'annotation/format/DS' to get the dosage.
note that not all gds files will have all nodes available, so choose an option that is compatible with the gds_file supplied.
}
